database update

+++ watchlist/commands.py
@@ -9,7 +9,9 @@ from watchlist.models import User, Movie
 def initdb(drop):
     """Initialize the database."""
     if drop:
+        click.confirm('This operation will delete the database, do you want to continue?', abort=True)
         db.drop_all()
+        click.echo('Drop tables.')
     db.create_all()
     click.echo('Initialized database.')
